refactor: organize agent prompts into dedicated folder
Move all agent system prompts into `components/agent/prompts/` for better organization and separation of concerns.

## Changes
- Create `components/agent/prompts/` directory
- Move prompt.ts, rag-prompt.ts, and web-scraper-prompt.ts into prompts/
- Update import paths in API routes:
  - app/api/chat/route.ts
  - app/api/rag-agent/route.ts
  - app/api/agent-with-mcp-tools/route.ts
- Update CLAUDE.md documentation to reflect new structure
