+++ app/api/agent-with-mcp-tools/route.ts
@@ -1,4 +1,4 @@
-import { WEB_SCRAPER_SYSTEM_INSTRUCTIONS } from "@/components/agent/web-scraper-prompt";
+import { WEB_SCRAPER_SYSTEM_INSTRUCTIONS } from "@/components/agent/prompts/web-scraper-prompt";
 import { getFirecrawlMCPClient } from "@/lib/mcp";
 import { openai } from "@ai-sdk/openai";
 import { streamText, convertToModelMessages, stepCountIs } from "ai";
