+++ app/api/chat/route.ts
@@ -1,4 +1,4 @@
-import { SYSTEM_INSTRUCTIONS } from "@/components/agent/prompt";
+import { SYSTEM_INSTRUCTIONS } from "@/components/agent/prompts/prompt";
 import { openai } from "@ai-sdk/openai";
 import { streamText, convertToModelMessages } from "ai";
 import { NextRequest } from "next/server";
